Issue 5210: Fixed Networking unset and side menu toggle
This bug was arising due to a JS error: "Uncaught TypeError: Cannot call method 'append' of null".
I changed the call from $(document).ready to $("body").load since the body tag was not getting defined by the time $(document).ready was called 
and hence it returned append of null.
By changing it to body.load we can achieve the same. I have tested it on FireFox and Chrome and the side menu states are not lost 
when Networking tab is opened.