fix's encoding error content in search pages
when click on search link in a forum page or with search form in sidebar
display an encoding error in web browser.
search.php?search_within[]=forums === error encoding content blank page.
I think after search in code, It's an oversights of removing [] in links
search_within with [] = page error, with not = page ok